LCDP/LCDPBackend/requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
djangorestframework-simplejwt>=5.3.0
psycopg2-binary>=2.9.0
python-decouple>=3.8

# PSO and NLP Dependencies
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
transformers>=4.21.0
torch>=1.12.0
spacy>=3.4.0
openai>=0.27.0
jieba>=0.42.1
nltk>=3.8
pandas>=1.5.0
matplotlib>=3.6.0 